Implement Async
The method .runCommandAsync from the Player and Entity Classes execute commands at the end of the tick while allowing the rest of the code to continue. Truth be told, we want such commands to execute before proceeding forward so we wait. Async defines the asynchronus function which returns a Promise while await is used to pause that function being executed until the Promise is fulfilled. In this patch we are essentially telling many functions to hold up until the method .runCommandAsync is done.