Restore uniqueId to maintain the integrity of permissions
The uniqueId is held in memory. Under a specific condition verify if this uniqueId still exists and recreate it if it is lost. An example that can result in a loss is to reload the server. In such cases, the memory is wiped. Since the hash is stored as a dynamic property we can validate this against the player and then regenerate the key and value so permissions are properly retained.