Optimize getPrefix
We use a const declaration for tags since we won't be modifying it. We also declare customprefix with an initial value of null instead of undefined for clarity.

We use for...of loop syntax to iterate over the tags array, and we use a break statement to exit the loop as soon as we find a tag that starts with "Prefix:".

Instead of returning config.customcommands.prefix directly when a custom prefix isn't found, we use the logical OR (||) operator to set config.customcommands.prefix to customprefix if it's truthy (i.e. it has a value other than null or undefined).