Enhance antispam
We check if a player is sending 3 or more messages within a span of 2 seconds. If they are then we cancel their most recent message being sent and warned them that they are spamming. We count each warning as a strike. Three strikes or more and the player is banned from the world. Testing this new code has shown that you really have to be excessively sending messages to be banned so if it happens to a player then its a pretty sure bet they had it coming to them.