Update antispam to reduce strike over time
Added a new property lastOffenseTime to the ChatRecord interface to keep track of the time of the last offense, and added a new else if block to check if the difference between the current time and the lastOffenseTime is greater than or equal to strikeReset, and if so, decrement the offenseCount by 1. For every minute that passes the strike on the player will be reduced by 1 to give them the chance to redeem themselves and eliminate the possibility of being banned based upon their behavior and disciplinary abilities to adhere to the rules.