[TPR] command to send teleport requests
This is code for handling teleportation requests (TPR) in a Minecraft server. The code defines an interface for teleport requests and uses an array to store these requests. There are two functions that handle these requests - one for submitting requests and one for approving/denying them. Additionally, the code also defines a function that sends usage instructions and a help message. The teleport requests are approved or denied based on the response received in the chat from the target player. The expiration time of the requests is also configurable using the configuration file. Default setting for the expiration time is set to two minutes.