[credits] Add Pete9xi and Frost
As Pete9xi has continued to maintain and support the project, it would be respectful to include him as a contributor and recognize him as a Developer to the project.

We would also like to acknowledge Frost for his contribution to the project. During the time when I was away, he worked on a forked repo and kept it alive for a while. His dedication to updating Paradox with each new release from Molang was instrumental in keeping it alive, and provided the initial momentum to keep the project going.