feat: Implement timer map to track player grace period for tp
This commit adds functionality to track a timer for each player using a Map data structure. The setTimer function sets a timer for 2 seconds for the given player and stores it in the map. The isTimerExpired function checks if the timer for a player has expired or not and automatically deletes it if it has.

This will allow players to safely teleport using the teleporting methods Paradox offers while not being flagged for speed hacking. This will not protect a player from being flagged for speed hacking if using other teleport methods outside of Paradox.

In addition, the commit defines the maximum age of a key and an interval for checking keys. The system runs an interval function to check if a key in the map has not been updated for a certain amount of time and should be removed to save memory. The implementation uses a lastUpdated property to track when a key was last updated and check if it has been longer than the desired time frame.