[worldborder] Refactor portal block detection logic
We first generate an array of block coordinates using destructuring assignment. Then, we loop through the coordinates and use a dictionary to store the block types. Finally, we check if the portal block or air block is in the dictionary using portalBlocks[MinecraftBlockTypes.portal.id] and portalBlocks[${x},${y-1},${z}] === MinecraftBlockTypes.air.id, respectively.