Policing key-value pairs in maps
startTimer is a function that starts a timer for a given key-value pair in a map called theVoid. The purpose of this function is to help manage memory by automatically deleting keys from the map after a specified period of time.

In programming, it's common to use maps to store key-value pairs, which can be useful for storing temporary data or caching information. However, if these keys are not deleted after they are no longer needed, they can start to accumulate and take up unnecessary memory. This is where startTimer comes in: by setting a timer for each key-value pair, the function ensures that keys are automatically deleted after a certain amount of time has passed.