[clearlag] Correction made to countdown
In the else if statement that checks if the time left is less than 60000 milliseconds, I added an if statement to check if the time left is exactly 60000 milliseconds. If it is, the function will send a message to the players saying "Server lag will be cleared in 1 minute!". If the time left is not exactly 60000 milliseconds, the function will calculate the number of seconds left, and send a message to the players saying "Server lag will be cleared in X seconds!", where X is the number of seconds left in the countdown. I also updated the message to use the singular form of the word "second" if there is only one second left in the countdown.