[lockdown] Enable command in config
This was disabled in e3aa0816aa8efb80b73057e0a033d0f1500408ec. This was due to Dynamic Properties being implemented into Paradox but shortly later discovering that DP had issues with server and realms upon being released with the API. This issue has been fixed for a very long time now and is no longer a concern so we can turn the command back on.