[registry] Fix "if (!flag)" block of code
This code will execute the if (!flag) block for each element in the defineBooleanProperties array that does not match with the config.modules object, and create a new dynamic property with a default value of false. Before this patch it was erroneously executing once and after the loop was complete.