[chatfilter] Improve performance and logic
In the optimized code, we are using a more performant way of iterating over tags, using a for...of loop instead of forEach, which avoids creating a new function for each iteration. Additionally, we are avoiding unnecessary string operations like replaceAll() by using replace() instead.

Furthermore, we are caching the sendMsg() function to avoid repeated lookups of the function in the module hierarchy. We are also using destructuring to directly access the properties of msg instead of accessing them through the msg object, which can be slightly faster.