[unban] Add new command to delete names from the list
The "delete" argument is one of the optional arguments that can be passed to the "unban" command. When "delete" is passed as an argument, the code attempts to remove a specified player's name from the "queueUnban" set using the "delete" method. If the name exists in the set, it is removed, and a message is sent to all players with the "paradoxOpped" tag to notify them that the player has been removed from the unban queue. If the name does not exist in the set, a message is sent to the command sender informing them that the player is not in the unban queue.