[illegalitems A, B, C] Use array to ignore certain items
there are certain items that need to be ignored because their data is not part of the API anymore. The list of items to be ignored is specified in the uniqueItems array.

This code checks if the current item being checked is not in the uniqueItems array, and if it is salvageable and meets the criteria for salvage, then the item will be salvaged.