[namespoofb] Fix TypeError
TypeError: regexp must have the 'g' flag at replaceAll

This error is occurring because the replaceAll() method requires a regular expression with the global flag (g) in order to replace all occurrences of the pattern in the string.

The replaceAll() method was introduced in ES2021 as an alternative to replace() that does not require the use of a regular expression with the global g flag. However, in the code, the regular expression we are passing to replaceAll() has a mixture of regular expression flags and special characters that are not allowed. Specifically, we are using the ^ and $ anchors, which are only valid when used with the m flag (multiline mode).