Set specific modules to false by default
These pertain to illegal items. Mojang has patched the game rather well to prevent this from being a problem. The code pertaining to Illegalitems A, B, and C as well as illegal lores, illegal enchantments, and illegal stacks need to be refactored and optimized dramatically and until this is accomplished it would be suggested to leave these off for now since they are not detrimental at this time.