[ops] Eliminate possible infinite loop
In the queueSleep function, the system.clearRun(id) call is used to cancel the interval that was set with system.runInterval. However, in the ops function, the id variable is reused inside the for loop to store the return value of system.runInterval, which creates a new interval every time the loop runs.

Therefore, when the queueSleep function is called and tries to cancel the interval with system.clearRun(id), it will only cancel the last interval that was created inside the for loop, leaving the previous intervals running. This will result in multiple instances of the queueSleep function running concurrently, causing the "Good Morning" message to be displayed continuously.

To fix this we eliminated the intervals and their corresponding ID's. Instead we take a more subtle approach by counting 2 seconds using milliseconds and once that has expired then we call the queueSleep function one time. To take this subtle approach one step further we only target the first player found in the array to handle this logic since it is not necessary to loop through each player and iterate over the same code per player. This is useless, redundant, and a waste of time and resources.

All in all, the updated ops() function should fix the infinite loop issue because it no longer uses a for loop to wait for the specified time. Instead, it uses Date.now() to measure the elapsed time since the start of the loop, and waits until it has reached the desired duration before calling queueSleep().