Delete Reach B and rename Reach C to B
Reach B was used to monitor broken blocks in a world. If broken beyond a distance not allowed it would restore that block to its original state then remove the dropped items. At face value this worked but when using a tool with the Fortune Enchantment the numbe of items dropped from the broken block increases. The code failed to remove the additional items properly. After many attempts at targeting all of the items it was decided to remove this feature entirely since we could not successfully prevent any form of duplications due to the results of the enchantment.