[clearlag] Clear old ID when reloading
When you reload the script while logged in the server, the previous interval doesn't appear to clear properly, and a new interval is started with a new ID. This results in multiple intervals temporarily running simultaneously, causing a jittery countdown behavior.

To play it safe and handle this situation delicately we initially set clearLagId to null. Inside the ClearLag() function, we first check if clearLagId is not null. If it's not null, we use system.clearRun() to cancel the previous interval with the given clearLagId.

Then, we start a new interval and assign its ID to clearLagId. The clearLagId is passed as an argument to clearLag() function to ensure that it receives the correct ID.

With this change, when you reload the script, the previous interval will be properly cleared before starting a new one, ensuring that only one interval is running at a time and preventing the jittery countdown behavior.