[worldborder] Bug fix for nether and safety protocol
The code was erroneously identfying the nether dimension due to a
in the string.

The safety protocol is used to teleport a player to a safe location
ideally with an air block to prevent being trapped or suffocated.

This function has been updated to get the relative position of the
players feet, body, and head to verify when all three instances return
air then safely teleport the player.