Bundle minecraft/vanilla-data with Paradox
For now it seems that the developers have designed this to be required. Its not an actual module but a placeholder for Types and other things to minimize the bloat in the minecraft/server module. Nobody knows at this time if bundling this with the behavior pack will remain a requirement in the future or if this is temporary. For now we ship it.