Update command handler for new Before and After events
The before event essentially listens to verify if a command is given using the prefix and cancels messages being sent if true or if the prefix command doesn't exist we cancel the messages and notify the player. If in cases where a prefix is recognised we listen to the after event where we handle the commands.

Canceling the message in the before event now cancels the after event since there is no message being broadcasted any longer. Canceling should only be used in delicate cases where its suitable. Otherwise we simply prevent broadcasting to any targeted player by ignoring them using 'sendToTargets'. This effectively allows the command handler to continue functioning by design in 1.20 with the current changes made in the API without causing issues, overlapping, race conditions, etc.