Update hasEnchantment()
vanilla-data returns an enum for MinecraftEnchantmentTypes. In the "for loop" it returns a key with each iteration. The code plugged it in assuming it was returning the string. We now correctly get the string by using the key as an index for MinecraftEnchantmentTypes to return the actual string for hasEnchantment() so we can obtain the proper number in return.