[antispam] Split between before and after chatSend events
This commit refactors the chat spam prevention functionality by moving the spam prevention logic to the "before" event handler, allowing early detection of spamming behavior. Then handling the banning functionality in the "after" event for proper, safe, and permissible execution of removing the player. Included the "extractTagInfo" function to assist on parsing through the object sent to the "after" event to handle the player in question of violating the rules for spamming.