entityHit -> entityHitEntity
In the original entityHit events all entities and blocks were included and as such either cases would trigger this. The community requested that entities and blocks be separated as this was creating too much overhead. In doing so they created entityHitEntity and entityHitBlock. This patch makes the neceassary changes to bring the specific scripts up to par with the changes found in entityHitEntity.