[AntiSpam] Fix ban reports of N/A
For some time now users have reported Paradox banning players with "By" and "Reason" showing N/A and during that time it couldn't be determined what was causing it. Today that problem is resolved after recieving a ss of one potential victim and seeing their list of tags. That list quickly directed me here and it was then when I realized the fatal mistake made that has been muysteriously taunting many. When defining the cause and effect it was not done so accurately.