[listitems] Resolve type error in enum iteration
The previous code caused a type error due to an implicit 'any' type. This commit fixes the issue by adding a type assertion to explicitly specify the type of the 'item' variable as 'keyof typeof MinecraftItemTypes'.

By doing this, we ensure that the 'item' variable is correctly used to index the 'MinecraftItemTypes' enum, allowing the string value to be retrieved without any type errors.

This change resolves the error and allows for the successful creation of ItemStack instances using the enum values.